edits made to flask app file
